Fix cache corruption, temp file leaks, and other issues from code review
- Fix removeIndices comparing write-position instead of read-position,
  which misaligned the download cache index when pruning more than one
  entry in a single pass (both the v1->v2 cache migration and the
  empty-entry cleanup run on every install/update/export/rehash).
- Stop leaking temp files under <cache>/temp/: every failure path in
  downloadNewFile and CreateFromTemp now closes and removes the temp
  file instead of abandoning it.
- Fix CacheIndexHandle.Remove() only deleting entries from hash-format
  lists the handle happened to have a value for, which could desync
  sparse format lists from the rest of the cache index.
- Parallelize StartDownloads() with a bounded worker pool instead of
  downloading one file at a time, guarding the shared CacheIndex with
  a mutex around the read-modify-write sections.
- Fix GetValidMCVersions sorting oldest-to-newest despite the comment
  (and callers) expecting newest-to-oldest.
- Guard the unchecked *v.Primary (and one *v.Filename) dereferences on
  nullable Modrinth API fields behind a nil-safe isPrimary() helper.
- Make PackFile.FileSize a uint64 so files >= 4GiB aren't silently
  truncated in the exported .mrpack manifest.
- Make getBestHash's last-resort fallback pick deterministically
  instead of relying on Go's randomised map iteration order.

Adds regression tests for each fix, including a -race-checked
concurrency test for the download worker pool.

+++ modrinth/export.go
@@ -160,7 +160,7 @@ var exportCmd = &cobra.Command{
 						Server string `json:"server"`
 					}{Client: clientEnv, Server: serverEnv},
 					Downloads: []string{u},
-					FileSize:  uint32(fileSize),
+					FileSize:  fileSize,
 				})
 
 				fmt.Printf("%s (%s) added to manifest\n", dl.Mod.Name, dl.Mod.FileName)
